+++ lambdas/thumbnail/CHANGELOG.md
@@ -1,3 +1,4 @@
+<!-- markdownlint-disable line-length -->
 # Changelog
 
 Changes are listed in reverse chronological order (newer entries at the top).
@@ -16,6 +17,7 @@ where verb is one of
 
 ## Changes
 
+- [Fixed] Clean up tmp directory to deal with tmp files persisted between invocations when lambda is killed because of OOM ([#4627](https://github.com/quiltdata/quilt/pull/4627))
 - [Fixed] Fix handling of some .tiff files ([#4609](https://github.com/quiltdata/quilt/pull/4609))
 - [Fixed] Fix handling of .czi files ([#4609](https://github.com/quiltdata/quilt/pull/4609))
 - [Changed] Switch from aicsimageio to bioio ([#4609](https://github.com/quiltdata/quilt/pull/4609))